requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.8.2
black==20.8b1
blessings==1.7
bpython==0.20
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
curtsies==0.3.4
Django==3.1.2
django-extensions==3.0.9
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.12.1
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
graphviz==0.14.2
greenlet==0.4.17
html5lib==1.1
idna==2.10
jdcal==1.4.1
l18n==2018.5
mypy-extensions==0.4.3
openpyxl==3.0.5
pathspec==0.8.0
Pillow==7.2.0
pydotplus==2.0.2
Pygments==2.7.2
pyparsing==2.4.7
pytz==2020.1
regex==2020.10.28
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==2.0.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
Unidecode==1.1.1
urllib3==1.25.11
wagtail==2.10.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
Willow==1.4
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
